VB-2693 added SQS config to project and basic SQS code to test if devops has been correctly deploy, removed blank lines after review
