--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsnotificationsalertsvsip/HmppsNotificationsAlertsVsip.kt
@@ -1,11 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication()
-class HmppsNotificationsAlertsVsip
-
-fun main(args: Array<String>) {
-  runApplication<HmppsNotificationsAlertsVsip>(*args)
-}
