--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsnotificationsalertsvsip/config/ResourceServerConfiguration.kt
@@ -1,49 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.config
-
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.cache.CacheManager
-import org.springframework.cache.annotation.EnableCaching
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.annotation.web.invoke
-import org.springframework.security.config.http.SessionCreationPolicy
-import org.springframework.security.oauth2.jwt.JwtDecoder
-import org.springframework.security.oauth2.jwt.NimbusJwtDecoder
-import org.springframework.security.web.SecurityFilterChain
-
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity(prePostEnabled = true, proxyTargetClass = true)
-@EnableCaching
-class ResourceServerConfiguration {
-
-  @Throws(Exception::class)
-  @Bean
-  fun filterChain(http: HttpSecurity): SecurityFilterChain {
-    http {
-      headers { frameOptions { sameOrigin = true } }
-      sessionManagement { sessionCreationPolicy = SessionCreationPolicy.STATELESS }
-      // Can't have CSRF protection as requires session
-      csrf { disable() }
-      authorizeHttpRequests {
-        listOf(
-          "/health/**",
-          "/info",
-          "/v3/api-docs/**",
-        ).forEach { authorize(it, permitAll) }
-        authorize(anyRequest, authenticated)
-      }
-      oauth2ResourceServer { jwt { jwtAuthenticationConverter = AuthAwareTokenConverter() } }
-    }
-    return http.build()
-  }
-
-  @Bean
-  fun locallyCachedJwtDecoder(
-    @Value("\${spring.security.oauth2.resourceserver.jwt.jwk-set-uri}") jwkSetUri: String,
-    cacheManager: CacheManager,
-  ): JwtDecoder = NimbusJwtDecoder.withJwkSetUri(jwkSetUri).cache(cacheManager.getCache("jwks")).build()
-}
