+++ src/main/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/HmppsNotificationsAlertsVsip.kt
@@ -1,0 +1,11 @@
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication()
+class HmppsNotificationsAlertsVsip
+
+fun main(args: Array<String>) {
+  runApplication<HmppsNotificationsAlertsVsip>(*args)
+}
