+++ src/main/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/service/listeners/events/EventFeatureSwitch.kt
@@ -1,0 +1,11 @@
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip.service.listeners.events
+
+import org.springframework.core.env.Environment
+import org.springframework.stereotype.Component
+
+@Component
+class EventFeatureSwitch(private val environment: Environment) {
+  fun isEnabled(eventType: String): Boolean {
+    return environment.getProperty("feature.event.$eventType", Boolean::class.java, true)
+  }
+}
