+++ src/main/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/service/listeners/events/SQSMessage.kt
@@ -1,0 +1,13 @@
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip.service.listeners.events
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+@Suppress("PropertyName")
+data class SQSMessage(
+  @JsonProperty("Type")
+  val type: String,
+  @JsonProperty("Message")
+  val message: String,
+  @JsonProperty("MessageId")
+  val messageId: String? = null,
+)
