--- src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsnotificationsalertsvsip/integration/domainevents/EventsIntegrationTestBase.kt
@@ -1,149 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.integration.domainevents
-
-import com.fasterxml.jackson.core.json.JsonWriteFeature
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.extension.ExtendWith
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.boot.test.mock.mockito.SpyBean
-import org.springframework.http.HttpHeaders
-import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.context.DynamicPropertyRegistry
-import org.springframework.test.context.DynamicPropertySource
-import software.amazon.awssdk.services.sns.model.PublishRequest
-import software.amazon.awssdk.services.sqs.SqsAsyncClient
-import software.amazon.awssdk.services.sqs.model.PurgeQueueRequest
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.helper.JwtAuthHelper
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.integration.domainevents.LocalStackContainer.setLocalStackProperties
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.integration.mock.HmppsAuthExtension
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.DomainEventListenerService
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.PRISON_VISITS_NOTIFICATION_ALERTS_QUEUE_CONFIG_KEY
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.listeners.events.DomainEvent
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.listeners.events.EventFeatureSwitch
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.listeners.events.SQSMessage
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.listeners.notifiers.PrisonVisitBookedEventNotifier
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.listeners.notifiers.PrisonVisitCancelledEventNotifier
-import uk.gov.justice.digital.hmpps.hmppsnotificationsalertsvsip.service.listeners.notifiers.PrisonVisitChangedEventNotifier
-import uk.gov.justice.hmpps.sqs.HmppsQueue
-import uk.gov.justice.hmpps.sqs.HmppsQueueService
-import uk.gov.justice.hmpps.sqs.HmppsTopic
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-@ExtendWith(HmppsAuthExtension::class)
-abstract class EventsIntegrationTestBase {
-
-  companion object {
-    private val localStackContainer = LocalStackContainer.instance
-    val objectMapper: ObjectMapper = ObjectMapper().registerModule(JavaTimeModule())
-      .disable(JsonWriteFeature.QUOTE_FIELD_NAMES.mappedFeature())
-
-    @JvmStatic
-    @DynamicPropertySource
-    fun testcontainers(registry: DynamicPropertyRegistry) {
-      localStackContainer?.also { setLocalStackProperties(it, registry) }
-    }
-
-    @BeforeAll
-    @JvmStatic
-    fun startMocks() {
-    }
-
-    @AfterAll
-    @JvmStatic
-    fun stopMocks() {
-    }
-  }
-
-  @Autowired
-  lateinit var domainEventListenerService: DomainEventListenerService
-
-  @SpyBean
-  lateinit var eventFeatureSwitch: EventFeatureSwitch
-
-  @SpyBean
-  lateinit var prisonVisitBookedEventNotifierSpy: PrisonVisitBookedEventNotifier
-
-  @SpyBean
-  lateinit var prisonVisitChangedEventNotifierSpy: PrisonVisitChangedEventNotifier
-
-  @SpyBean
-  lateinit var prisonVisitCancelledEventNotifierSpy: PrisonVisitCancelledEventNotifier
-
-  @Autowired
-  protected lateinit var objectMapper: ObjectMapper
-
-  @Autowired
-  private lateinit var hmppsQueueService: HmppsQueueService
-
-  internal val topic by lazy { hmppsQueueService.findByTopicId("domainevents") as HmppsTopic }
-
-  internal val prisonVisitsEventsQueue by lazy { hmppsQueueService.findByQueueId(PRISON_VISITS_NOTIFICATION_ALERTS_QUEUE_CONFIG_KEY) as HmppsQueue }
-
-  internal val sqsClient by lazy { prisonVisitsEventsQueue.sqsClient }
-  internal val sqsDlqClient by lazy { prisonVisitsEventsQueue.sqsDlqClient }
-  internal val queueUrl by lazy { prisonVisitsEventsQueue.queueUrl }
-  internal val dlqUrl by lazy { prisonVisitsEventsQueue.dlqUrl }
-
-  internal val awsSnsClient by lazy { topic.snsClient }
-  internal val topicArn by lazy { topic.arn }
-
-  lateinit var roleVisitSchedulerHttpHeaders: (HttpHeaders) -> Unit
-
-  @Autowired
-  protected lateinit var jwtAuthHelper: JwtAuthHelper
-
-  @BeforeEach
-  fun cleanQueue() {
-    purgeQueue(sqsClient, queueUrl)
-    purgeQueue(sqsDlqClient!!, dlqUrl!!)
-    roleVisitSchedulerHttpHeaders = setAuthorisation(roles = listOf("ROLE_VISIT_SCHEDULER"))
-  }
-
-  internal fun setAuthorisation(
-    user: String = "AUTH_ADM",
-    roles: List<String> = listOf(),
-    scopes: List<String> = listOf(),
-  ): (HttpHeaders) -> Unit = jwtAuthHelper.setAuthorisation(user, roles, scopes)
-
-  fun purgeQueue(client: SqsAsyncClient, url: String) {
-    client.purgeQueue(PurgeQueueRequest.builder().queueUrl(url).build()).get()
-  }
-
-  fun createDomainEvent(eventType: String, additionalInformation: String = "test"): DomainEvent {
-    return DomainEvent(eventType = eventType, additionalInformation)
-  }
-
-  fun createSQSMessage(domainEventJson: String): String {
-    val sqaMessage = SQSMessage(type = "Notification", messageId = "123", message = domainEventJson)
-    return objectMapper.writeValueAsString(sqaMessage)
-  }
-
-  fun createDomainEventPublishRequest(eventType: String, domainEvent: String): PublishRequest? {
-    return PublishRequest.builder()
-      .topicArn(topicArn)
-      .message(domainEvent).build()
-  }
-
-  fun createDomainEventPublishRequest(eventType: String): PublishRequest? {
-    return PublishRequest.builder()
-      .topicArn(topicArn)
-      .message(objectMapper.writeValueAsString(createDomainEvent(eventType, ""))).build()
-  }
-
-  fun createDomainEventJson(eventType: String, additionalInformation: String): String {
-    return "{\"eventType\":\"$eventType\",\"additionalInformation\":$additionalInformation}"
-  }
-
-  fun createAdditionalInformationJson(effectiveDate: String, expiryDate: String? = null): String {
-    val builder = StringBuilder()
-    builder.append("{")
-    // TODO Must be in JSON format
-    builder.append("}")
-    return builder.toString()
-  }
-}
