+++ src/test/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/health/HealthInfoTest.kt
@@ -1,0 +1,16 @@
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip.health
+
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Test
+import org.springframework.boot.info.BuildProperties
+import java.util.Properties
+
+class HealthInfoTest {
+  @Test
+  fun `should include version info`() {
+    val properties = Properties()
+    properties.setProperty("version", "somever")
+    Assertions.assertThat(HealthInfo(BuildProperties(properties)).health().details)
+      .isEqualTo(mapOf("version" to "somever"))
+  }
+}
