+++ src/test/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/integration/IntegrationTestBase.kt
@@ -1,0 +1,16 @@
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip.integration
+
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.web.reactive.server.WebTestClient
+
+@SpringBootTest(webEnvironment = RANDOM_PORT)
+@ActiveProfiles("test")
+abstract class IntegrationTestBase {
+
+  @Suppress("SpringJavaInjectionPointsAutowiringInspection")
+  @Autowired
+  lateinit var webTestClient: WebTestClient
+}
